fix for skip and top

+++ src/Avalara.SDK/Api/EInvoicing/V1/DataInputFieldsApi.cs
@@ -32,9 +32,9 @@ namespace Avalara.SDK.Api.EInvoicing.V1
         /// </summary>
         public int? Top { get; set; }
         /// <summary>
-        /// If nonzero, skip this number of results before returning data. Used with &lt;code&gt;$top&lt;/code&gt; to provide pagination for large datasets.
+        /// The number of items to skip in the result.
         /// </summary>
-        public decimal? Skip { get; set; }
+        public int? Skip { get; set; }
         /// <summary>
         /// When set to true, the count of the collection is also returned in the response body
         /// </summary>
